requirements.txt
ultralytics>=8.0.0
opencv-python>=4.7.0
pytesseract>=0.3.10
gTTS>=2.3.2
playsound>=1.3.0
openai>=0.27.0
fastapi>=0.95.0
uvicorn>=0.21.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
numpy>=1.24.0
Pillow>=9.5.0
